reorganized files and folder structure, added cobra
