+++ cloudhunter/go.mod
@@ -1,0 +1,15 @@
+module github.com/Kimi99/cloudhunter
+
+go 1.24.4
+
+require github.com/aws/aws-sdk-go v1.55.7
+
+require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.6 // indirect
+)
+
+require (
+	github.com/jmespath/go-jmespath v0.4.0 // indirect
+	github.com/spf13/cobra v1.9.1
+)
