+++ cloudhunter/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/Kimi99/cloudhunter/cmd"
+
+func main() {
+	cmd.Execute()
+}
