--- cloudhunter/src/go.mod
@@ -1,7 +1,0 @@
-module github.com/Kimi99/cloudhunter
-
-go 1.24.4
-
-require github.com/aws/aws-sdk-go v1.55.7
-
-require github.com/jmespath/go-jmespath v0.4.0 // indirect
